Move parsing collapsed format to separate file in utils directory
